refactor: Separate controller directory
