+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import baseball.domain.BaseballGame;
+import baseball.controller.BaseballGame;
 
 public class Application {
     public static void main(String[] args) {
