+++ src/main/java/baseball/model/Player.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.model;
 
 import baseball.status.GameStatus;
