+++ src/test/java/baseball/model/PlayerTest.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.model;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
